% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignTR.R
\name{alignTRs}
\alias{alignTRs}
\title{Aligning Tandem Repeats}
\usage{
alignTRs(encoded_trs)
}
\arguments{
\item{encoded_trs}{A vector of encoded tandem repeats}
}
\value{
A character matrix representing the aligned tandem repeats, where
each row corresponds to an input sequence and each column to an aligned
motif position.
}
\description{
Aligns multiple encoded tandem repeat sequences by motif using a simple
progressive pairwise alignment algorithm
}
\examples{
\dontrun{
#Example 1
encoded_trs <- c("AABA", "ABA", "BA")
alignTRs(encoded_trs)

###expected
#' [,1] [,2] [,3] [,4]
TR1 "A"  "A"  "B"  "A"
TR2 "-"  "A"  "B"  "A"
TR3 "-"  "-"  "B"  "A"
}

}
\references{
OpenAI. ChatGPT (GPT-5) large language model (2025).
https://chat.openai.com/
}
