% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeTR.R
\name{decomposeTRs}
\alias{decomposeTRs}
\title{Decomposition of Tandem Repeats}
\usage{
decomposeTRs(alleles, motifs, match_score = 1, indel = -1, allowance = 0)
}
\arguments{
\item{alleles}{A \code{DNAStringSet} or character vector of tandem repeat alleles.}

\item{motifs}{A vector of strings or a \code{DNAStringSet} representing the
expected motifs of the tandem repeat.}

\item{match_score}{Numeric. Score assigned when a motif matches the sequence.
Default is \code{1}.}

\item{indel}{Numeric. Score assigned for an insertion or deletion in the
sequence. Default is \code{-1}.}

\item{allowance}{Integer. Number of mismatches allowed between the motif and
the sequence for it to still be encoded as that motif. Default is \code{0}.}
}
\value{
A list containing:
\item{\strong{compositions}}{ The tandem repeat alleles decomposed into their
constituent motifs. If an allele contains only one tandem repeat, this
element is returned as a vector. Otherwise, it is a list of vectors.}
\item{\strong{motifs}}{ A \code{DNAStringSet} of motifs including any motifs present
across all supplied alleles.}

If no motifs are found in an allele, \code{compositions} will contain
\code{character(0)} for that allele, and a message is printed to inform the user.
}
\description{
Decomposes alleles into their constituent motifs given either a character
vector or a \code{DNAStringSet} of alleles.
}
\examples{
\dontrun{
##Example 1

library(Biostrings)
alleles <- DNAStringSet(c("TTTACACGTAC", "ACACGTAC", "TTACACGTAC" ))
motifs <- DNAStringSet(c("AC", "GT"))
compositions <- decomposeTRs(alleles, motifs)
compositions

###Expected:
$compositions
$compositions[[1]]
[1] "TTT" "AC"  "AC"  "GT"  "AC"

$compositions[[2]]
[1] "AC" "AC" "GT" "AC"

$compositions[[3]]
[1] "TT" "AC" "AC" "GT" "AC"


$motifs
DNAStringSet object of length 4:
  width seq
[1]     2 AC
[2]     2 GT
[3]     3 TTT
[4]     2 TT
}

\dontrun{
##Example 2

allele <- DNAString("TTTACACGTCAC")
motifs <- DNAStringSet(c("AC", "GTC"))
composition <- decomposeTRs(allele, motifs)
composition

###Expected
$compositions
[1] "TTT" "AC" "AC" "GTC" "AC"
$motifs
DNAStringSet object of length 3:
  width seq
[1]     2 AC
[2]     3 GTC
[3]     3 TTT
}

}
\references{
Pagès, H., Aboyoun, P., Gentleman, R. & DebRoy, S. Biostrings: Efficient
manipulation of biological strings (R package version 2.77.2, 2025).
https://bioconductor.org/packages/Biostrings, doi:10.18129/B9.bioc.Biostrings

Park, J., Kaufman, E., Valdmanis, P. N. & Bafna, V. TRviz: A Python Library
for decomposing and Visualizing Tandem Repeat Sequences. Bioinformatics
Advances 3, (2023).

R Core Team. R: A Language and Environment for Statistical Computing. Vienna,
Austria: R Foundation for Statistical Computing (2025).
https://www.R-project.org/

OpenAI. ChatGPT (GPT-5) large language model (2025).
https://chat.openai.com/
}
