% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_palette}
\alias{default_palette}
\title{Default TRskim color palette (colorblind-safe)}
\format{
A character vector of hex color codes.
}
\usage{
default_palette
}
\description{
A vector of 66 distinct colors used by \code{plotTR()}
when no palette is supplied.
}
\keyword{datasets}
