% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeTR.R
\name{encodeTR}
\alias{encodeTR}
\title{Tandem Repeat Encoder}
\usage{
encodeTR(decomposed_TR, motifs)
}
\arguments{
\item{decomposed_TR}{A character vector containing the decomposed tandem
repeat (ex. "-"  "-"  "AC" "AC" "GT" "AC")}

\item{motifs}{A character vector containing the motifs that may be in the TR}
}
\value{
Returns a list containing the encoded tandem repeat in $encoded and the
mapping of the motifs to their symbols in $motif_map
}
\description{
A function that encodes a decomposed tandem repeat such that each unique
motif is a single character symbol
}
\examples{
\dontrun{

decomposed_TR <- c("AC", "AC", "GT", "AC")
motifs <- c("AC", "GT")
encoded <- encodeTR(decomposed_TR, motifs)

encoded$encoded
[1] "AABA"

encoded$map
 AC  GT
"A" "B"

}

}
\references{
Park, J., Kaufman, E., Valdmanis, P. N. & Bafna, V. TRviz: A Python Library
for decomposing and Visualizing Tandem Repeat Sequences. Bioinformatics
Advances 3, (2023).

R Core Team. R: A Language and Environment for Statistical Computing.
(R Foundation for Statistical Computing, 2025). Available at:
https://www.R-project.org/
}
