% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motifs_SORL1}
\alias{motifs_SORL1}
\title{DNAStringSet containing 36 Motifs}
\format{
A DNAStringSet with 36 motif sequences
}
\source{
TRviz example data
https://github.com/Jong-hun-Park/trviz/blob/main/examples/sample_code.ipynb

Park, J., Kaufman, E., Valdmanis, P. N. & Bafna, V. TRviz: A Python Library
for decomposing and Visualizing Tandem Repeat Sequences. Bioinformatics
Advances 3, (2023).
}
\usage{
motifs_SORL1
}
\description{
DNAStringSet containing 36 motifs that are in the tandem repeat SORL1
}
\examples{
\dontrun{
motifs_SORL1}

}
\keyword{datasets}
