% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Running The TRskim Workflow}
\usage{
run_workflow(alleles, motifs, match_score = 1, indel = -1, allowance = 0)
}
\arguments{
\item{alleles}{DNAStringSet or character string vector of the Tandem repeats}

\item{motifs}{a vector of strings or a DNAStringSet that are the expected
motifs of the Tandem Repeat}

\item{match_score}{score for if motif matches sequence this is by default 1}

\item{indel}{score for if there is an insertion or deletion in sequence by
default this is -1}

\item{allowance}{is the number of mismatches between the motif
and the sequence permitted to be still be later encoded as a motif. By
default this is 0}
}
\value{
The function returns a list containing the decompositions
($decomposition), encodings ($encoded), motif map ($motif_map)
alignment matrix ($alignment), bar plot visualization ($bar_plot) and tile
plot visualization ($tile_plot) of the TRskim package work flow.
}
\description{
A function that performs all the steps of the TRskim workflow
}
\references{
Pagès, H., Aboyoun, P., Gentleman, R. & DebRoy, S. Biostrings: Efficient
manipulation of biological strings (R package version 2.77.2, 2025).
https://bioconductor.org/packages/Biostrings, doi:10.18129/B9.bioc.Biostrings
}
