% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTRskim.R
\name{runTRskim}
\alias{runTRskim}
\title{Launch Shiny App for TRskim}
\usage{
runTRskim()
}
\value{
launches the shiny app
}
\description{
This functions launches the Shiny app for TRskim. The app is intended for
those with less coding experience and performs the entire TRskim workflow.
}
\examples{
\dontrun{
TRskim::runTRskim()
}

}
