% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SORL1}
\alias{SORL1}
\title{SOL1 tandem Repeat alleles}
\format{
A DNAStringSet with 95 alleles
}
\source{
TRviz example data
https://github.com/Jong-hun-Park/trviz/tree/main/examples/data

Park, J., Kaufman, E., Valdmanis, P. N. & Bafna, V. TRviz: A Python Library
for decomposing and Visualizing Tandem Repeat Sequences. Bioinformatics
Advances 3, (2023).
}
\usage{
SORL1
}
\description{
DNASet containing 95 SORL1 tandem repeat alleles.
}
\examples{
\dontrun{
SORL1}


}
\keyword{datasets}
