% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeTR.R
\name{decomposeTR}
\alias{decomposeTR}
\title{Decomposes allele into motifs}
\usage{
decomposeTR(allele, motifs, match = 1, indel = -1, allowence = 0)
}
\arguments{
\item{allele}{DNAString or character string of the Tandem repeat}

\item{motifs}{a vector of strings or a DNAStringSet that are the expected
motifs of the Tandem Repeat}

\item{match}{score for if motif matches sequence this is by default 1}

\item{indel}{score for if there is an insertion or deletion in sequence by
default this is -1}

\item{allowence}{is the number of mismatches between the motif
and the sequence permitted to be still be later encoded as a motif.By default
this is 0}
}
\description{
A function that decomposes an allele into the supplied motifs given either a
character string or DNAString of the allele
}
\references{
OpenAI. (2025). ChatGPT (GPT-5) \link{Large language model}. https://chat.openai.com/
}
