% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeTR.R
\name{encodeTRs}
\alias{encodeTRs}
\title{Encoding Tandem Repeats}
\usage{
encodeTRs(decomposed_TRs, motifs)
}
\arguments{
\item{decomposed_TRs}{A list or character vector containing the decomposed
tandem repeat(s) (e.g., \code{c("AC", "AC", "GT", "AC")}).}

\item{motifs}{A character vector or \code{DNAStringSet} containing the motifs that
may appear in the tandem repeats.}
}
\value{
A list with two elements:
\item{encoded}{The encoded tandem repeats. Returned as a list if there are
multiple tandem repeats, or as a single string if only one repeat is
provided.}
\item{motif_map}{A named vector mapping motifs to their corresponding
single-character symbols.}
}
\description{
Encodes decomposed tandem repeats such that each unique motif is represented
by a single character symbol. \strong{Note:} Maximum of 66 unique encodings is
allowed.
}
\examples{
\dontrun{
decomposed_TR1 <- c("TTT", "AC", "AC", "GT", "AC")
decomposed_TR2 <- c("AC", "GT", "AC")
decomposed_TR3 <- c("CC", "AC", "GT", "AC", "GT")
decomposed_TRs <- list(decomposed_TR1, decomposed_TR2, decomposed_TR3)
motifs <- c("AC", "GT", "TTT", "CC")
encoded <- encodeTRs(decomposed_TRs, motifs)
encoded

$encoded
$encoded[[1]]
[1] "CAABA"

$encoded[[2]]
[1] "ABA"

$encoded[[3]]
[1] "DABAB"


$motif_map
AC  GT TTT  CC
"A" "B" "C" "D"
}

\dontrun{

decomposed_TR <- c("AC", "AC", "GT", "AC")
motifs <- c("AC", "GT")
encoded <- encodeTRs(decomposed_TR, motifs)
encoded

$encoded
[1] "AABA"

$motif_map
 AC  GT
"A" "B"

}
}
\references{
OpenAI. ChatGPT (GPT-5) large language model (2025).https://chat.openai.com/

Park, J., Kaufman, E., Valdmanis, P. N. & Bafna, V. TRviz: A Python Library
for decomposing and Visualizing Tandem Repeat Sequences. Bioinformatics
Advances 3, (2023).

R Core Team. R: A Language and Environment for Statistical Computing.
(R Foundation for Statistical Computing, 2025). Available at:
https://www.R-project.org/
}
