% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeTR.R
\name{encodeTRs}
\alias{encodeTRs}
\title{Encoding Tandem Repeats}
\usage{
encodeTRs(decomposed_TRs, motifs)
}
\arguments{
\item{decomposed_TRs}{A list or chracter vector containing the decomposed
tandem repeat(s) (ex."AC" "AC" "GT" "AC")}

\item{motifs}{A character vector containing the motifs that may be
in the TR}
}
\value{
Returns encoded tandem repeats in $encoded either as a list if there is
multiple tandem repeats or as a single string if there is only one and the
mapping of the motifs to their symbols in $motif_map
}
\description{
A function that encodes decomposed tandem repeats such that each unique
motif is a single character symbol. \strong{MAXIMUM OF 66 ENCODINGS}
}
\examples{
\dontrun{
decomposed_TR1 <- c("TTT", "AC", "AC", "GT", "AC")
decomposed_TR2 <- c("AC", "GT", "AC")
decomposed_TR3 <- c("CC", "AC", "GT", "AC", "GT")
decomposed_TRs <- list(decomposed_TR1, decomposed_TR2, decomposed_TR3)
motifs <- c("AC", "GT", "TTT", "CC")
encoded <- encodeTRs(decomposed_TRs, motifs)
encoded
$encoded
$encoded[[1]]
[1] "CAABA"

$encoded[[2]]
[1] "ABA"

$encoded[[3]]
[1] "DABAB"


$motif_map
AC  GT TTT  CC
"A" "B" "C" "D"
}

\dontrun{

decomposed_TR <- c("AC", "AC", "GT", "AC")
motifs <- c("AC", "GT")
encoded <- encodeTRs(decomposed_TR, motifs)

encoded$encoded
[1] "AABA"

encoded$map
 AC  GT
"A" "B"

}
}
\references{
OpenAI. ChatGPT (GPT-5) large language model (2025).https://chat.openai.com/

Park, J., Kaufman, E., Valdmanis, P. N. & Bafna, V. TRviz: A Python Library
for decomposing and Visualizing Tandem Repeat Sequences. Bioinformatics
Advances 3, (2023).

R Core Team. R: A Language and Environment for Statistical Computing.
(R Foundation for Statistical Computing, 2025). Available at:
https://www.R-project.org/
}
